+++ src/main/kotlin/minesweeper/Minesweeper.kt
@@ -1,16 +1,20 @@
 package minesweeper
 
-import minesweeper.domain.Height
-import minesweeper.domain.MineField
-import minesweeper.domain.Width
+import minesweeper.domain.field.Height
+import minesweeper.domain.field.MineField
+import minesweeper.domain.field.Width
 import minesweeper.view.InputView
 
 object Minesweeper {
 
     fun start() {
+        val field = createField()
+        InputView.mineCount
+    }
+
+    private fun createField(): MineField {
         val height = InputView.height.let(::Height)
         val width = InputView.width.let(::Width)
-        MineField.of(height, width)
-        InputView.mineCount
+        return MineField.of(height, width)
     }
 }
