+++ src/main/kotlin/minesweeper/domain/cell/Cell.kt
@@ -1,4 +1,4 @@
-package minesweeper.domain
+package minesweeper.domain.cell
 
 data class Cell(
     val row: Int,
