+++ src/main/kotlin/minesweeper/domain/field/Width.kt
@@ -1,4 +1,4 @@
-package minesweeper.domain
+package minesweeper.domain.field
 
 data class Width(
     val value: Int
