+++ src/test/kotlin/domain/cell/CellTest.kt
@@ -1,9 +1,9 @@
-package domain
+package domain.cell
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
-import minesweeper.domain.Cell
-import minesweeper.domain.CellMark
+import minesweeper.domain.cell.Cell
+import minesweeper.domain.cell.CellMark
 
 class CellTest : StringSpec({
     "셀 생성시 지뢰가 없다" {
